requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
langchain==0.2.14
langchain-openai==0.1.22
pinecone-client==5.0.1
sentence-transformers==3.1.1
ragas==0.1.11
pandas==2.2.2
scikit-learn==1.5.1
pypdf==4.3.1
